Are you sure
git-svn-id: https://fluid.surelogic.com/svn/eng/trunk/sierra-client-eclipse@3918 d02ae147-ee12-4a35-9452-91c89e8b2592

+++ src/com/surelogic/sierra/client/eclipse/views/SierraServersMediator.java
@@ -68,7 +68,12 @@ public final class SierraServersMediator implements ISierraServerObserver {
 		public void handleEvent(Event event) {
 			SierraServer server = f_manager.getFocus();
 			if (server != null) {
-				f_manager.delete(server);
+				if (MessageDialog.openConfirm(event.display.getActiveShell(),
+						"Confirm Sierra Server Deletion",
+						"Do you wish to delete the Sierra server '"
+								+ server.getLabel() + "'?")) {
+					f_manager.delete(server);
+				}
 			}
 		}
 	};
